guide db js6 with template
