+++ src/guide/1/dbsearch/javascript/snippets/create.test.yml
@@ -3,7 +3,7 @@ description: Load SDK
 hooks:
   before:
   after:
-template: empty
+template: catch
 expected: New documents successfully created!
 sdk: js
 version: 6
