+++ src/guide/1/dbsearch/javascript/snippets/final.test.yml
@@ -1,0 +1,11 @@
+---
+name: dbsearch#search
+description: Create Index collection and documents and apply a search request.
+hooks:
+  before:
+  after:
+template: 
+expected: There is 1 document that match.
+
+sdk: js
+version: 6
